BREAKING CHANGE: #86 useSyncMorfeo
