+++ packages/hooks/src/index.ts
@@ -1,5 +1,5 @@
 export * from './useTheme';
 export * from './useProps';
 export * from './useStyles';
-export * from './useMorfeo';
+export * from './useSyncMorfeo';
 export * from './useCurrentTheme';
