+++ packages/hooks/src/useCurrentTheme.ts
@@ -1,9 +1,9 @@
-import { ThemeName } from '@morfeo/core';
-import { useMorfeo } from './useMorfeo';
+import { morfeo, ThemeName } from '@morfeo/core';
+import { useSyncMorfeo } from './useSyncMorfeo';
 
 type UseCurrentThemeReturnValue = [ThemeName, (name: ThemeName) => void];
 
 export function useCurrentTheme(): UseCurrentThemeReturnValue {
-  const { getCurrentTheme, setCurrentTheme } = useMorfeo();
-  return [getCurrentTheme(), setCurrentTheme];
+  useSyncMorfeo();
+  return [morfeo.getCurrentTheme(), morfeo.setCurrentTheme];
 }
