+++ packages/react/src/useClassName.ts
@@ -1,11 +1,11 @@
-import { useState, useEffect } from 'react';
+import { useState, useInsertionEffect } from 'react';
 import stringify from 'fast-safe-stringify';
 import { Style, getStyles } from '@morfeo/web';
 
 export function useClassNames<Key extends string>(styles: Record<Key, Style>) {
   const [{ update, classes }] = useState(() => getStyles(styles));
 
-  useEffect(() => {
+  useInsertionEffect(() => {
     update(styles);
   }, [stringify(styles)]);
